HHH-13041 - Move @Any and other association mappings to the association chapter
