+++ documentation/src/main/asciidoc/userguide/chapters/domain/extras/associations/associations-JoinColumnOrFormula-fetching-example.sql
