+++ documentation/src/main/asciidoc/userguide/chapters/domain/extras/associations/associations-JoinFormula-example.sql
