+++ documentation/src/main/asciidoc/userguide/chapters/domain/extras/associations/associations-any-query-example.sql
