+++ documentation/src/main/asciidoc/userguide/chapters/domain/extras/associations/associations-many-to-any-example.sql
