+++ documentation/src/main/asciidoc/userguide/chapters/domain/extras/embeddable/embeddable-Target-fetching-example.sql
