+++ documentation/src/main/asciidoc/userguide/chapters/pc/extras/pc-filter-entity-example.sql
@@ -1,0 +1,13 @@
+SELECT
+    a.id as id1_0_0_,
+    a.active_status as active2_0_0_,
+    a.amount as amount3_0_0_,
+    a.client_id as client_i6_0_0_,
+    a.rate as rate4_0_0_,
+    a.account_type as account_5_0_0_,
+    c.id as id1_1_1_,
+    c.name as name2_1_1_ 
+FROM
+    Account a
+WHERE
+    a.id = 2
