+++ documentation/src/main/asciidoc/userguide/chapters/pc/extras/pc-filter-join-table-persistence-example.sql
@@ -1,0 +1,23 @@
+INSERT INTO Client (name, id)
+VALUES ('John Doe', 1)
+
+INSERT INTO Account (amount, client_id, rate, account_type, id)
+VALUES (5000.0, 1, 0.0125, 'CREDIT', 1)
+
+INSERT INTO Account (amount, client_id, rate, account_type, id)
+VALUES (0.0, 1, 0.0105, 'DEBIT', 2)
+
+INSERT INTO Account (amount, client_id, rate, account_type, id)
+VALUES (250.0, 1, 0.0105, 'DEBIT', 3)
+
+INSERT INTO Client_Account (Client_id, order_id, accounts_id)
+VALUES (1, 0, 1)
+
+INSERT INTO Client_Account (Client_id, order_id, accounts_id)
+VALUES (1, 0, 1)
+
+INSERT INTO Client_Account (Client_id, order_id, accounts_id)
+VALUES (1, 1, 2)
+
+INSERT INTO Client_Account (Client_id, order_id, accounts_id)
+VALUES (1, 2, 3)
