+++ documentation/src/main/asciidoc/userguide/chapters/pc/extras/pc-filter-persistence-example.sql
@@ -1,0 +1,11 @@
+INSERT INTO Client (name, id)
+VALUES ('John Doe', 1)
+
+INSERT INTO Account (active_status, amount, client_id, rate, account_type, id)
+VALUES (true, 5000.0, 1, 0.0125, 'CREDIT', 1)
+
+INSERT INTO Account (active_status, amount, client_id, rate, account_type, id)
+VALUES (false, 0.0, 1, 0.0105, 'DEBIT', 2)
+
+INSERT INTO Account (active_status, amount, client_id, rate, account_type, id)
+VALUES (true, 250.0, 1, 0.0105, 'DEBIT', 3)
