+++ documentation/src/test/java/org/hibernate/userguide/associations/JoinColumnOrFormulaTest.java
@@ -4,19 +4,21 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.userguide.mapping.basic;
+package org.hibernate.userguide.associations;
 
+import java.io.Serializable;
 import java.util.Objects;
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
+import org.hibernate.annotations.JoinColumnOrFormula;
 import org.hibernate.annotations.JoinFormula;
-import org.hibernate.dialect.PostgreSQL82Dialect;
 import org.hibernate.jpa.test.BaseEntityManagerFunctionalTestCase;
 
-import org.hibernate.testing.RequiresDialect;
 import org.junit.Test;
 
 import org.jboss.logging.Logger;
@@ -27,8 +29,7 @@ import static org.junit.Assert.assertEquals;
 /**
  * @author Vlad Mihalcea
  */
-@RequiresDialect(PostgreSQL82Dialect.class)
-public class JoinFormulaTest extends BaseEntityManagerFunctionalTestCase {
+public class JoinColumnOrFormulaTest extends BaseEntityManagerFunctionalTestCase {
 
 	@Override
 	protected Class<?>[] getAnnotatedClasses() {
@@ -40,14 +41,18 @@ public class JoinFormulaTest extends BaseEntityManagerFunctionalTestCase {
 
 	@Test
 	public void testLifecycle() {
-		//tag::mapping-JoinFormula-persistence-example[]
+		//tag::associations-JoinColumnOrFormula-persistence-example[]
 		Country US = new Country();
 		US.setId( 1 );
+		US.setDefault( true );
+		US.setPrimaryLanguage( "English" );
 		US.setName( "United States" );
 
 		Country Romania = new Country();
 		Romania.setId( 40 );
+		Romania.setDefault( true );
 		Romania.setName( "Romania" );
+		Romania.setPrimaryLanguage( "Romanian" );
 
 		doInJPA( this::entityManagerFactory, entityManager -> {
 			entityManager.persist( US );
@@ -59,19 +64,20 @@ public class JoinFormulaTest extends BaseEntityManagerFunctionalTestCase {
 			user1.setId( 1L );
 			user1.setFirstName( "John" );
 			user1.setLastName( "Doe" );
-			user1.setPhoneNumber( "+1-234-5678" );
+			user1.setLanguage( "English" );
 			entityManager.persist( user1 );
 
 			User user2 = new User( );
 			user2.setId( 2L );
 			user2.setFirstName( "Vlad" );
 			user2.setLastName( "Mihalcea" );
-			user2.setPhoneNumber( "+40-123-4567" );
+			user2.setLanguage( "Romanian" );
 			entityManager.persist( user2 );
-		} );
-		//end::mapping-JoinFormula-persistence-example[]
 
-		//tag::mapping-JoinFormula-fetching-example[]
+		} );
+		//end::associations-JoinColumnOrFormula-persistence-example[]
+
+		//tag::associations-JoinColumnOrFormula-fetching-example[]
 		doInJPA( this::entityManagerFactory, entityManager -> {
 			log.info( "Fetch User entities" );
 
@@ -81,10 +87,10 @@ public class JoinFormulaTest extends BaseEntityManagerFunctionalTestCase {
 			User vlad = entityManager.find( User.class, 2L );
 			assertEquals( Romania, vlad.getCountry());
 		} );
-		//end::mapping-JoinFormula-fetching-example[]
+		//end::associations-JoinColumnOrFormula-fetching-example[]
 	}
 
-	//tag::mapping-JoinFormula-example[]
+	//tag::associations-JoinColumnOrFormula-example[]
 	@Entity(name = "User")
 	@Table(name = "users")
 	public static class User {
@@ -96,15 +102,28 @@ public class JoinFormulaTest extends BaseEntityManagerFunctionalTestCase {
 
 		private String lastName;
 
-		private String phoneNumber;
+		private String language;
 
 		@ManyToOne
-		@JoinFormula( "REGEXP_REPLACE(phoneNumber, '\\+(\\d+)-.*', '\\1')::int" )
+		@JoinColumnOrFormula( column =
+			@JoinColumn(
+				name = "language",
+				referencedColumnName = "primaryLanguage",
+				insertable = false,
+				updatable = false
+			)
+		)
+		@JoinColumnOrFormula( formula =
+			@JoinFormula(
+				value = "true",
+				referencedColumnName = "is_default"
+			)
+		)
 		private Country country;
 
 		//Getters and setters omitted for brevity
 
-	//end::mapping-JoinFormula-example[]
+	//end::associations-JoinColumnOrFormula-example[]
 		public Long getId() {
 			return id;
 		}
@@ -129,36 +148,45 @@ public class JoinFormulaTest extends BaseEntityManagerFunctionalTestCase {
 			this.lastName = lastName;
 		}
 
-		public String getPhoneNumber() {
-			return phoneNumber;
+		public String getLanguage() {
+			return language;
 		}
 
-		public void setPhoneNumber(String phoneNumber) {
-			this.phoneNumber = phoneNumber;
+		public void setLanguage(String language) {
+			this.language = language;
 		}
 
 		public Country getCountry() {
 			return country;
 		}
 
-	//tag::mapping-JoinFormula-example[]
-	}
-	//end::mapping-JoinFormula-example[]
+		public void setCountry(Country country) {
+			this.country = country;
+		}
 
-	//tag::mapping-JoinFormula-example[]
+	//tag::associations-JoinColumnOrFormula-example[]
+	}
+	//end::associations-JoinColumnOrFormula-example[]
+
+	//tag::associations-JoinColumnOrFormula-example[]
 
 	@Entity(name = "Country")
 	@Table(name = "countries")
-	public static class Country {
+	public static class Country implements Serializable {
 
 		@Id
 		private Integer id;
 
 		private String name;
 
+		private String primaryLanguage;
+
+		@Column(name = "is_default")
+		private boolean _default;
+
 		//Getters and setters, equals and hashCode methods omitted for brevity
 
-	//end::mapping-JoinFormula-example[]
+	//end::associations-JoinColumnOrFormula-example[]
 
 		public int getId() {
 			return id;
@@ -174,6 +202,22 @@ public class JoinFormulaTest extends BaseEntityManagerFunctionalTestCase {
 
 		public void setName(String name) {
 			this.name = name;
+		}
+
+		public String getPrimaryLanguage() {
+			return primaryLanguage;
+		}
+
+		public void setPrimaryLanguage(String primaryLanguage) {
+			this.primaryLanguage = primaryLanguage;
+		}
+
+		public boolean isDefault() {
+			return _default;
+		}
+
+		public void setDefault(boolean _default) {
+			this._default = _default;
 		}
 
 		@Override
@@ -192,7 +236,7 @@ public class JoinFormulaTest extends BaseEntityManagerFunctionalTestCase {
 		public int hashCode() {
 			return Objects.hash( getId() );
 		}
-	//tag::mapping-JoinFormula-example[]
+	//tag::associations-JoinColumnOrFormula-example[]
 	}
-	//end::mapping-JoinFormula-example[]
+	//end::associations-JoinColumnOrFormula-example[]
 }
