+++ documentation/src/test/java/org/hibernate/userguide/associations/any/AnyTest.java
@@ -4,26 +4,25 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.userguide.mapping.basic.any;
+package org.hibernate.userguide.associations.any;
 
 import org.hibernate.testing.junit4.BaseCoreFunctionalTestCase;
 import org.junit.Test;
 
 import static org.hibernate.testing.transaction.TransactionUtil.doInHibernate;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
 
 /**
  * @author Vlad Mihalcea
  */
-public class ManyToAnyTest extends BaseCoreFunctionalTestCase {
+public class AnyTest extends BaseCoreFunctionalTestCase {
 
     @Override
     protected Class<?>[] getAnnotatedClasses() {
         return new Class<?>[] {
             IntegerProperty.class,
             StringProperty.class,
-            PropertyRepository.class
+            PropertyHolder.class
         };
     }
 
@@ -38,7 +37,7 @@ public class ManyToAnyTest extends BaseCoreFunctionalTestCase {
     public void test() {
 
         doInHibernate( this::sessionFactory, session -> {
-            //tag::mapping-column-many-to-any-persist-example[]
+            //tag::associations-any-persist-example[]
             IntegerProperty ageProperty = new IntegerProperty();
             ageProperty.setId( 1L );
             ageProperty.setName( "age" );
@@ -53,26 +52,21 @@ public class ManyToAnyTest extends BaseCoreFunctionalTestCase {
 
             session.persist( nameProperty );
 
-            PropertyRepository propertyRepository = new PropertyRepository();
-            propertyRepository.setId( 1L );
+            PropertyHolder namePropertyHolder = new PropertyHolder();
+            namePropertyHolder.setId( 1L );
+            namePropertyHolder.setProperty( nameProperty );
 
-            propertyRepository.getProperties().add( ageProperty );
-            propertyRepository.getProperties().add( nameProperty );
-
-            session.persist( propertyRepository );
-            //end::mapping-column-many-to-any-persist-example[]
+            session.persist( namePropertyHolder );
+            //end::associations-any-persist-example[]
         } );
 
         doInHibernate( this::sessionFactory, session -> {
-            //tag::mapping-column-many-to-any-query-example[]
-            PropertyRepository propertyRepository = session.get( PropertyRepository.class, 1L );
+            //tag::associations-any-query-example[]
+            PropertyHolder propertyHolder = session.get( PropertyHolder.class, 1L );
 
-            assertEquals(2, propertyRepository.getProperties().size());
-
-            for(Property property : propertyRepository.getProperties()) {
-                assertNotNull( property.getValue() );
-            }
-            //end::mapping-column-many-to-any-query-example[]
+            assertEquals("name", propertyHolder.getProperty().getName());
+            assertEquals("John Doe", propertyHolder.getProperty().getValue());
+            //end::associations-any-query-example[]
         } );
     }
